+++ test/fixtures/cookbooks/test/recipes/server.rb
@@ -27,8 +27,8 @@ end
 samba_share 'first_share' do
   comment 'exported share 1'
   path '/srv/export'
-  guest_ok 'no'
-  printable 'no'
+  guest_ok true
+  printable false
   write_list ['test_user_1']
   create_mask '0644'
   directory_mask '0775'
@@ -38,8 +38,8 @@ end
 samba_share 'second_share' do
   comment 'exported share 2'
   path '/srv/export_2'
-  guest_ok 'no'
-  printable 'no'
+  guest_ok false
+  printable false
   write_list ['test_user_2']
   create_mask '0644'
   directory_mask '0775'